only report once
